delete transit, add static bus routes
