+++ PennMobile/src/main/res/layout/fragment_main.xml
@@ -11,7 +11,7 @@
     style="@style/MainRelativeLayout"
     android:id="@+id/registrar_cont"
     app:layout_widthPercent="50%"
-    app:layout_heightPercent="25%">
+    app:layout_heightPercent="33%">
 
     <ImageView
         android:contentDescription="@string/registrar"
@@ -38,7 +38,7 @@
         android:layout_toRightOf="@id/registrar_cont"
         android:id="@+id/directory_cont"
         app:layout_widthPercent="50%"
-        app:layout_heightPercent="25%">
+        app:layout_heightPercent="33%">
 
         <ImageView
             android:contentDescription="@string/directory"
@@ -64,7 +64,7 @@
         style="@style/MainRelativeLayout"
         android:id="@+id/dining_cont"
         app:layout_widthPercent="50%"
-        app:layout_heightPercent="25%"
+        app:layout_heightPercent="33%"
         android:layout_below="@id/registrar_cont">
 
         <ImageView
@@ -91,14 +91,14 @@
         style="@style/MainRelativeLayout"
         android:layout_toRightOf="@id/dining_cont"
         android:layout_below="@id/registrar_cont"
-        android:id="@+id/transit_cont"
+        android:id="@+id/laundry_cont"
         app:layout_widthPercent="50%"
-        app:layout_heightPercent="25%">
+        app:layout_heightPercent="33%">
 
         <ImageView
-            android:contentDescription="@string/transit"
-            android:id="@+id/transit_img"
-            android:src="@drawable/ic_transit"
+            android:contentDescription="@string/laundry"
+            android:id="@+id/laundry_img"
+            android:src="@drawable/ic_laundry"
             style="@style/MainIcon"
             app:layout_widthPercent="100%"
             app:layout_heightPercent="60%"
@@ -106,9 +106,9 @@
             />
 
         <TextView
-            android:id="@+id/transit_button"
-            android:layout_below="@+id/transit_img"
-            android:text="@string/transit"
+            android:id="@+id/laundry_button"
+            android:layout_below="@+id/laundry_img"
+            android:text="@string/laundry"
             style="@style/MainText"
             app:layout_widthPercent="100%"
             app:layout_heightPercent="25%"/>
@@ -119,7 +119,7 @@
         style="@style/MainRelativeLayout"
         android:id="@+id/news_cont"
         app:layout_widthPercent="50%"
-        app:layout_heightPercent="25%"
+        app:layout_heightPercent="33%"
         android:layout_below="@id/dining_cont">
 
         <ImageView
@@ -148,7 +148,7 @@
         android:layout_below="@id/dining_cont"
         android:id="@+id/map_cont"
         app:layout_widthPercent="50%"
-        app:layout_heightPercent="25%">
+        app:layout_heightPercent="33%">
 
         <ImageView
             android:contentDescription="@string/map"
@@ -170,31 +170,4 @@
 
     </android.support.percent.PercentRelativeLayout>
 
-    <android.support.percent.PercentRelativeLayout
-        style="@style/MainRelativeLayout"
-        android:layout_below="@id/news_cont"
-        android:id="@+id/laundry_cont"
-        app:layout_widthPercent="50%"
-        app:layout_heightPercent="25%"
-        android:layout_centerHorizontal="true"
-        >
-
-        <ImageView
-            android:contentDescription="@string/laundry"
-            android:id="@+id/laundry_img"
-            android:src="@drawable/ic_laundry"
-            app:layout_widthPercent="100%"
-            app:layout_heightPercent="60%"
-            app:layout_marginTopPercent="10%"
-            style="@style/MainIcon" />
-
-        <TextView
-            android:id="@+id/laundry_button"
-            android:layout_below="@+id/laundry_img"
-            android:text="@string/laundry"
-            style="@style/MainText"
-            app:layout_widthPercent="100%"
-            app:layout_heightPercent="25%"/>
-
-    </android.support.percent.PercentRelativeLayout>
 </android.support.percent.PercentRelativeLayout>
