--- PennMobile/src/main/res/menu/transit.xml
@@ -1,10 +1,0 @@
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item android:id="@+id/transit_search"
-        android:icon="?attr/searchIcon"
-        android:hint="@string/transit_search"
-        app:actionViewClass="android.support.v7.widget.SearchView"
-        android:actionViewClass="android.support.v7.widget.SearchView"
-        app:showAsAction="ifRoom"
-        android:title="@string/search_title"/>
-</menu>
