+++ PennMobile/src/main/res/values/strings.xml
@@ -63,6 +63,8 @@
     </string-array>
     <string name="google_api_origin">origin</string>
     <string name="google_api_dest">destination</string>
+    <string name="no_permission_map">Permission not granted for using Google Map</string>
+    <string name="ask_permission_fail">Access of your location is required to use this feature.</string>
 
     <!-- Transit search -->
     <string name="starting_location_hint">Current location</string>
